+++ packages/side-nav/src/SideNavItem/index.ts
@@ -1,0 +1,2 @@
+export { default as SideNavItem } from './SideNavItem';
+export { default as CollapsedSideNavItem } from './CollapsedSideNavItem';
