+++ packages/side-nav/src/index.ts
@@ -1,7 +1,6 @@
 import SideNav, { sideNavSelector } from './SideNav';
-import SideNavItem from './SideNavItem';
+import { SideNavItem, CollapsedSideNavItem } from './SideNavItem';
 import SideNavGroup from './SideNavGroup';
-import CollapsedSideNavItem from './CollapsedSideNavItem';
 import { sideNavWidth, sideNavItemSidePadding } from './styles';
 
 export {
